Update Form Validation logic
